backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
transformers
torch
pydantic
python-multipart